+++ src/main/java/kr/borntorun/api/adapter/in/web/FeedController.java
@@ -43,7 +43,7 @@ public class FeedController {
 
   @Operation(summary = "피드 상세 보기", description = "특정 피드를 조회합니다.")
   @GetMapping(value = "{feedId}", produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<DetailFeedResponse> detail(@AuthUser TokenDetail my, @PathVariable final int feedId) {
+  public ResponseEntity<DetailFeedResponse> detail(@AuthUser TokenDetail my, @PathVariable final long feedId) {
     final Feed feed = feedProxy.searchDetail(my, feedId);
     feedProxy.increaseViewQty(feedId);
 
@@ -58,13 +58,13 @@ public class FeedController {
 
   @Operation(summary = "피드 삭제", description = "피드를 삭제합니다.")
   @DeleteMapping(value = "/{feedId}", produces = MediaType.APPLICATION_JSON_VALUE)
-  public void remove(@AuthUser TokenDetail my, @PathVariable final int feedId) {
+  public void remove(@AuthUser TokenDetail my, @PathVariable final long feedId) {
     feedProxy.remove(feedId, my);
   }
 
   @Operation(summary = "피드 수정", description = "피드를 수정합니다.")
   @PutMapping(value = "/{feedId}", produces = MediaType.APPLICATION_JSON_VALUE)
-  public void modify(@AuthUser TokenDetail my, @PathVariable final int feedId, @RequestBody @Valid final ModifyFeedRequest request) {
+  public void modify(@AuthUser TokenDetail my, @PathVariable final long feedId, @RequestBody @Valid final ModifyFeedRequest request) {
     feedProxy.modify(request, feedId);
   }
 
@@ -72,7 +72,7 @@ public class FeedController {
   @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
   public ResponseEntity<Page<SearchFeedResponse>> searchAll(@AuthUser TokenDetail my,
       @Valid @ModelAttribute SearchFeedRequest request,
-      @RequestParam(defaultValue = "0") int lastFeedId,
+      @RequestParam(defaultValue = "0") long lastFeedId,
       @RequestParam(defaultValue = "10") int size) {
     final Page<FeedCard> feedPage = feedProxy.searchAll(request, my, lastFeedId, PageRequest.of(0, size));
