+++ src/main/java/kr/borntorun/api/adapter/in/web/payload/CreateFeedRequest.java
@@ -8,7 +8,7 @@ import kr.borntorun.api.domain.constant.FeedAccessLevel;
 import kr.borntorun.api.domain.constant.FeedCategory;
 
 @Schema(description = "피드 작성")
-public record CreateFeedRequest(@Schema(description = "업로드한 이미지 식별자 리스트") List<Integer> imageIds,
+public record CreateFeedRequest(@Schema(description = "업로드한 이미지 식별자 리스트") List<Long> imageIds,
                                 @Schema(description = "내용") @NotNull String contents,
                                 @Schema(description = "피드 종류") @NotNull FeedCategory category,
                                 @Schema(description = "공개 범위") @NotNull FeedAccessLevel accessLevel) {}
