+++ src/main/java/kr/borntorun/api/adapter/in/web/payload/SearchFeedResponse.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "피드 목록 조회")
-public record SearchFeedResponse(@Schema(description = "글 번호") int id,
+public record SearchFeedResponse(@Schema(description = "글 번호") long id,
                                  @Schema(description = "이미지 uri 리스트") List<String> imageUris,
                                  @Schema(description = "본문") String contents,
                                  @Schema(description = "조회수") long viewQty,
