+++ src/main/java/kr/borntorun/api/adapter/in/web/proxy/FeedProxy.java
@@ -34,19 +34,19 @@ public class FeedProxy {
   private final FeedService feedService;
 
   @Cacheable(key = "'searchDetail: ' + #feedId + #my.id")
-  public Feed searchDetail(final TokenDetail my, final int feedId) {
+  public Feed searchDetail(final TokenDetail my, final long feedId) {
     SearchFeedDetailCommand command = feedConverter.toSearchFeedDetailCommand(feedId, my);
     return feedService.searchDetail(command);
   }
 
   @Cacheable(key = "#request == null ? 'searchAll: ' + #my.id + #pageable.pageSize : 'searchAll: ' + #my.id + #request.hashCode() + #pageable.pageSize")
-  public Page<FeedCard> searchAll(final SearchFeedRequest request, final TokenDetail my, final int lastFeedId, final Pageable pageable) {
+  public Page<FeedCard> searchAll(final SearchFeedRequest request, final TokenDetail my, final long lastFeedId, final Pageable pageable) {
     SearchAllFeedCommand command = feedConverter.toSearchAllFeedCommand(request, my, lastFeedId);
     return feedService.searchAll(command, pageable);
   }
 
   //  @DistributedLock(key = "'FeedView-'.concat(#id)", waitTime = 10L)
-  public void increaseViewQty(final int feedId) {
+  public void increaseViewQty(final long feedId) {
     feedService.increaseViewQty(feedId);
   }
 
@@ -57,13 +57,13 @@ public class FeedProxy {
   }
 
   @CacheEvict(allEntries = true)
-  public void remove(final int feedId, final TokenDetail my) {
+  public void remove(final long feedId, final TokenDetail my) {
     RemoveFeedCommand command = feedConverter.toRemoveFeedCommand(feedId, my);
     feedService.remove(command);
   }
 
   @CacheEvict(allEntries = true)
-  public void modify(final ModifyFeedRequest request, final int feedId) {
+  public void modify(final ModifyFeedRequest request, final long feedId) {
     ModifyFeedCommand command = feedConverter.toModifyFeedCommand(request, feedId);
     feedService.modify(command);
   }
