+++ src/main/java/kr/borntorun/api/adapter/in/web/proxy/RecommendationProxy.java
@@ -22,13 +22,13 @@ public class RecommendationProxy {
   private final RecommendationService recommendationService;
 
   @CacheEvict(allEntries = true, cacheNames = {"recommendation", "feed"})
-  public void create(final TokenDetail my, final RecommendationType recommendationType, final int contentId) {
+  public void create(final TokenDetail my, final RecommendationType recommendationType, final long contentId) {
     CreateRecommendationCommand command = recommendationConverter.toCreateRecommendationCommand(my.getId(), recommendationType, contentId);
     recommendationService.create(command);
   }
 
   @CacheEvict(allEntries = true, cacheNames = {"recommendation", "feed"})
-  public void remove(final TokenDetail my, final RecommendationType recommendationType, final int contentId) {
+  public void remove(final TokenDetail my, final RecommendationType recommendationType, final long contentId) {
     RemoveRecommendationCommand command = recommendationConverter.toRemoveRecommendationCommand(my.getId(), recommendationType, contentId);
     recommendationService.remove(command);
   }
