+++ src/main/java/kr/borntorun/api/adapter/in/web/proxy/YellowCardProxy.java
@@ -16,7 +16,7 @@ public class YellowCardProxy {
 
   private final YellowCardService yellowCardService;
 
-  public void create(final int myUserId, final CreateYellowCardRequest request) {
+  public void create(final long myUserId, final CreateYellowCardRequest request) {
     CreateYellowCardCommand command = yellowCardConverter.toCreateYellowCardCommand(request, myUserId);
     yellowCardService.create(command);
   }
