+++ src/main/java/kr/borntorun/api/config/aop/ActivityParticipationWriterValidationAspect.java
@@ -22,17 +22,17 @@ public class ActivityParticipationWriterValidationAspect {
   public void onCancelUser() {}
 
   @Before("onCancelUser() && args(my, participationId, ..)")
-  public void validateCancel(final TokenDetail my, final int participationId) throws IllegalArgumentException {
+  public void validateCancel(final TokenDetail my, final long participationId) throws IllegalArgumentException {
     validate(my.getId(), participationId);
   }
 
-  private void validate(final int userId, final int participationId) {
+  private void validate(final long userId, final long participationId) {
     if (!isValid(userId, participationId)) {
       throw new ForBiddenException("잘못된 접근입니다.");
     }
   }
 
-  private boolean isValid(final int userId, final int participationId) {
+  private boolean isValid(final long userId, final long participationId) {
     return activityParticipationRepository.findById(participationId)
         .orElseThrow(() -> new NotFoundException("참여를 하지 않은 상태입니다."))
         .getUserId() == userId;
