+++ src/main/java/kr/borntorun/api/config/aop/CommentWriterValidationAspect.java
@@ -26,22 +26,22 @@ public class CommentWriterValidationAspect {
   public void onDeleteUser() {}
 
   @Before("onModifyUser() && args(my, commentId, request, ..)")
-  public void validateModify(final TokenDetail my, final int commentId, final ModifyCommentRequest request) throws IllegalArgumentException {
+  public void validateModify(final TokenDetail my, final long commentId, final ModifyCommentRequest request) throws IllegalArgumentException {
     validate(my.getId(), commentId);
   }
 
   @Before("onDeleteUser() && args(my, commentId, ..)")
-  public void validateDelete(final TokenDetail my, final int commentId) throws IllegalArgumentException {
+  public void validateDelete(final TokenDetail my, final long commentId) throws IllegalArgumentException {
     validate(my.getId(), commentId);
   }
 
-  private void validate(final int userId, final int feedId) {
+  private void validate(final long userId, final long feedId) {
     if (!isValid(userId, feedId)) {
       throw new ForBiddenException("잘못된 접근입니다.");
     }
   }
 
-  private boolean isValid(final int userId, final int commentId) {
+  private boolean isValid(final long userId, final long commentId) {
     return componentRepository.findById(commentId)
         .orElseThrow(() -> new NotFoundException("댓글를 찾을 수 없습니다."))
         .getUserId() == userId;
